test: add tests for versions pkg
